feat: auto-set PIF from fee amounts per claim type, unify PIF badges
