+++ drizzle/meta/_journal.json
@@ -225,6 +225,20 @@
       "when": 1783345622194,
       "tag": "0031_normalize_yes_fees_confirmation_casing",
       "breakpoints": true
+    },
+    {
+      "idx": 32,
+      "version": "7",
+      "when": 1783356588846,
+      "tag": "0032_pif_auto_suggestion",
+      "breakpoints": true
+    },
+    {
+      "idx": 33,
+      "version": "7",
+      "when": 1783357639164,
+      "tag": "0033_fix_pif_auto_write_skip_condition",
+      "breakpoints": true
     }
   ]
 }
